cleanup for usePublisher
