+++ frontend/src/utils/getAccessDeniedError/index.ts
@@ -1,0 +1,3 @@
+import getAccessDeniedError from './getAccessDeniedError';
+
+export default getAccessDeniedError;
